requirements.txt
streamlit
openai
langchain
pymilvus
python-dotenv
sentence_transformers
tiktoken
pypdf
